- double arithmetic - volume error down to 6‰
git-svn-id: https://svn.vis.uni-stuttgart.de/projects/megamol/plugins/mmstd_trisoup/trunk@45 cb50be13-817d-0410-ab5c-3777bfb244a9

+++ volumetrics/TetraVoxelizer.h
@@ -23,10 +23,10 @@ namespace volumetrics {
 
 		void growSurfaceFromTriangle(FatVoxel *theVolume, unsigned int x, unsigned int y, unsigned int z,
 							 unsigned char triIndex, 
-							 vislib::Array<float> &surf, vislib::Array<BorderVoxel *> &border,
-                             float &surfSurf, float &volume);
+							 vislib::Array<double> &surf, vislib::Array<BorderVoxel *> &border,
+                             double &surfSurf, double &volume);
 
-        float growVolume(FatVoxel *theVolume, unsigned int x, unsigned int y, unsigned int z);
+        double growVolume(FatVoxel *theVolume, unsigned int x, unsigned int y, unsigned int z);
 
 		bool CellHasNoGeometry(FatVoxel *theVolume, unsigned x, unsigned y, unsigned z);
